flask_server/requirements.txt
flask
flask-sqlalchemy
pymysql
cryptography
flask_wtf
wtforms
email_validator
flask_bcrypt
flask_login
psycopg2
mysql-connector-python